+++ packages/simple-taro-fast-framework/src/customConfig/constants.js
@@ -58,4 +58,13 @@ export const pathCollection = {
   notification: {
     path: '/example/pages/notification/index',
   },
+  grid: {
+    path: '/example/pages/grid/index',
+  },
+  selector: {
+    path: '/example/pages/selector/index',
+  },
+  tabs: {
+    path: '/example/pages/tabs/index',
+  },
 };