+++ packages/taro-fast-component/config/rollupConfig.js
@@ -1,9 +1,6 @@
 import { buildConfig as buildConfigCore } from '../../taro-fast-rollup/rollupAssist/configBuilder';
 
 const inputFile = {
-  // ...{
-  //   index: 'src/index.ts',
-  // },
   ...{
     'customComponents/index': 'src/customComponents/index.jsx',
   },