+++ packages/taro-fast-common/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [1.0.12](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-common@1.0.11...taro-fast-common@1.0.12) (2022-01-11)
+
+**Note:** Version bump only for package taro-fast-common
+
+
+
+
+
+### [1.0.11](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-common@1.0.10...taro-fast-common@1.0.11) (2022-01-11)
+
+**Note:** Version bump only for package taro-fast-common
+
+
+
+
+
+### [1.0.10](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-common@1.0.9...taro-fast-common@1.0.10) (2022-01-11)
+
+**Note:** Version bump only for package taro-fast-common
+
+
+
+
+
+### [1.0.9](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-common@1.0.8...taro-fast-common@1.0.9) (2022-01-11)
+
+**Note:** Version bump only for package taro-fast-common
+
+
+
+
+
+### [1.0.8](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-common@1.0.7...taro-fast-common@1.0.8) (2022-01-11)
+
+**Note:** Version bump only for package taro-fast-common
+
+
+
+
+
+### [1.0.7](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-common@1.0.6...taro-fast-common@1.0.7) (2022-01-11)
+
+**Note:** Version bump only for package taro-fast-common
+
+
+
+
+
+### [1.0.6](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-common@1.0.5...taro-fast-common@1.0.6) (2022-01-11)
+
+**Note:** Version bump only for package taro-fast-common
+
+
+
+
+
 ### [1.0.5](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-common@1.0.4...taro-fast-common@1.0.5) (2022-01-02)
 
 **Note:** Version bump only for package taro-fast-common